# Fix denoise_nl_means data type management