--- a/skimage/_shared/fast_exp.h
+++ b/skimage/_shared/fast_exp.h
@@ -3,14 +3,17 @@
  * Reference [2]: https://doi.org/10.1162/089976600300015033
  * Additional improvements by Leonid Bloch. */
 
+#include <stdint.h>
+
 /* use just EXP_A = 1512775 for integer version, to avoid FP calculations */
-#define EXP_A (1512775.3951951856938)  /* 2^20*ln2 */
+#define EXP_A (1512775.3951951856938)  /* 2^20÷ln2 */
 /* For min. RMS error */
 #define EXP_BC 1072632447              /* 1023*2^20 - 60801 */
 /* For min. max. relative error */
 /* #define EXP_BC 1072647449 */        /* 1023*2^20 - 45799 */
 /* For min. mean relative error */
 /* #define EXP_BC 1072625005 */        /* 1023*2^20 - 68243 */
+
 
 __inline double fast_exp (double y)
 {
@@ -32,6 +35,39 @@ __inline double fast_exp (double y)
         case 0:
             /* Big endian */
             _eco.n.i = (int)(EXP_A*(y)) + EXP_BC;
+            _eco.n.j = 0;
+            break;
+    }
+
+    return _eco.d;
+}
+
+
+#define EXPF_A (184.6649652337873)  /* 2^7÷ln2 */
+/* For min. RMS error */
+#define EXPF_BC 16249              /* 127*2^7 - 7 */
+
+
+__inline float fast_expf (float y)
+{
+    union
+    {
+        double d;
+        struct { int16_t i, j; } n;
+        char t[8];
+    } _eco;
+
+    _eco.n.i = 1;
+
+    switch(_eco.t[0]) {
+        case 1:
+            /* Little endian */
+            _eco.n.j = (int)(EXPF_A*(y)) + EXPF_BC;
+            _eco.n.i = 0;
+            break;
+        case 0:
+            /* Big endian */
+            _eco.n.i = (int)(EXPF_A*(y)) + EXPF_BC;
             _eco.n.j = 0;
             break;
     }