--- a/skimage/_shared/fast_exp.pyx
+++ b/skimage/_shared/fast_exp.pyx
@@ -1,6 +1,6 @@
 from .fused_numerics cimport np_floats
-from .fast_exp cimport exp_func
+from .fast_exp cimport _fast_exp
 
 
-def exp_func_py(np_floats x):
-    return exp_func(x)
+def fast_exp(np_floats x):
+    return _fast_exp(x)