--- a/skimage/_shared/setup.py
+++ b/skimage/_shared/setup.py
@@ -15,12 +15,14 @@ def configuration(parent_package='', top_path=None):
 
     cython(['geometry.pyx',
             'transform.pyx',
-            'interpolation.pyx'], working_path=base_path)
+            'interpolation.pyx',
+            'fast_exp.pyx'], working_path=base_path)
 
     config.add_extension('geometry', sources=['geometry.c'])
     config.add_extension('transform', sources=['transform.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('interpolation', sources=['interpolation.c'])
+    config.add_extension('fast_exp', sources=['fast_exp.c'])
     return config