… benchmarks/benchmark_restoration.py
… skimage/_shared/tests/test_fast_exp.py
… skimage/restoration/setup.py
… skimage/restoration/tests/test_denoise.py