@rfezzani 2019-12-03
Here are the results for the demo example @grlee77, every thing looks fine ;-)

## Double precision result

![nl_means](https://user-images.githubusercontent.com/3438227/70033855-936c5800-15b0-11ea-8c8d-a7b01bffe34a.png)

```python
>>> print(f"PSNR (noisy) = {psnr_noisy:0.2f}")
PSNR (noisy) = 22.16
>>> print(f"PSNR (slow) = {psnr:0.2f}")
PSNR (slow) = 29.38
>>> print(f"PSNR (slow, using sigma) = {psnr2:0.2f}")
PSNR (slow, using sigma) = 29.77
>>> print(f"PSNR (fast) = {psnr_fast:0.2f}")
PSNR (fast) = 28.43
>>> print(f"PSNR (fast, using sigma) = {psnr2_fast:0.2f}")
PSNR (fast, using sigma) = 29.26
```
## Single precision result
![nl_means_f32](https://user-images.githubusercontent.com/3438227/70033867-99facf80-15b0-11ea-8317-088f5d209275.png)

```python
>>> print(f"PSNR (noisy) = {psnr_noisy:0.2f}")
PSNR (noisy) = 22.16
>>> print(f"PSNR (slow) = {psnr:0.2f}")
PSNR (slow) = 29.38
>>> print(f"PSNR (slow, using sigma) = {psnr2:0.2f}")
PSNR (slow, using sigma) = 29.77
>>> print(f"PSNR (fast) = {psnr_fast:0.2f}")
PSNR (fast) = 28.44
>>> print(f"PSNR (fast, using sigma) = {psnr2_fast:0.2f}")
PSNR (fast, using sigma) = 29.26
```