@rfezzani 2019-12-04
I propose to add a note about the possible artifacts apparition due to round-off errors accumulation when single precision is used on large volumetric data. Is it OK for you?