@grlee77 2019-12-19
> I propose to add a note about the possible artifacts apparition due to round-off errors accumulation when single precision is used on large volumetric data. Is it OK for you?

I would be okay with it, but it seems that a more conservative and backwards compatible solution is to just always leave the `integral` arrays as double rather than allowing fused types there. This wouldn't prevent using float32 for the input & output arrays.