@rfezzani 2020-02-10
@grlee77, I fixed the `integral` array type to `double`, and removed the note about round off error accumulation when using single precision. There is no more backward compatibility issue. Do you see any other issue with this PR?