@grlee77 2020-02-11
@jni:
 I just tried the basic benchmark pasted below (involving 256 x 256 x 32 volumetric data) and this PR is a bit faster than master.

|  Branch  | Precision | Duration
| -------- | --------- | --------
| master  | double | 4.27 s 
| master  | single | 4.39 s
| ----------- |  ----------| ----------
| this PR | double | 4.04 s
| this PR | single | 3.59 s

```Python
from time import time
import numpy as np
import skimage
from skimage.restoration import denoise_nl_means
import skimage.data

nz = 32
vol = np.stack([skimage.data.camera()[::2, ::2], ] * nz, axis=-1).astype(float) / 255
sigma = .05
voln = vol + sigma * np.random.randn(*vol.shape)
volnf = voln.astype(np.float32)

tstart = time()
denoise_nl_means(voln, patch_size=3, patch_distance=2, sigma=sigma, h=0.7*sigma, fast_mode=True)
dur_double = time() - tstart
print(dur_double)

tstart = time()
denoise_nl_means(volnf, patch_size=3, patch_distance=2, sigma=sigma, h=0.7*sigma, fast_mode=True)
dur_single = time() - tstart
print(dur_single)
```