@rfezzani 2020-02-11
Thank you @grlee77 for tackling this, I complitely forgot about it (sorry @jni)