@jni 2020-02-11
@grlee77 that's great news!

@rfezzani it would be great to add the above to our benchmarks in this PR. Have you seen @hmaarrfk's benchmarks writing guide? https://scikit-image.org/docs/dev/contribute.html#benchmarks