@pep8speaks 2020-02-11
Hello @rfezzani! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`benchmarks/benchmark_restoration.py`](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py):

> [Line 22:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L22): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 27:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L27): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 32:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L32): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 37:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L37): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 42:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L42): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 47:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L47): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 52:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L52): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator
> [Line 57:49](https://github.com/scikit-image/scikit-image/blob/7e5034d35207be00a16849fcebd1dec473c79a74/benchmarks/benchmark_restoration.py#L57): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator